This patch fixes two bugs in the workflow::case::action::notify proc in the workflow package.  I am not very familiar with the way this package works so someone should check my work before releasing this into the wild.
Patch 1 - Retrieve the non-cached version of the activity log.  While
debugging my application I noticed that if I fired off two actions quite
quickly the second action's notification would be the same as the first.

Patch 2 - When firing an action that has no assignees the default
behavior is to notify EVERYONE who has subscribed to be notified of
their cases.  Since no one has been assigned it's nobody's case yet.
With this change no one will be notified except people who are
subscribed to the workflow if an action occurs on a case without any
assignees.